Rearranges and moves some methods from the acoustic harmonic solver.
