+++ vibra/engine/solvers/acoustic_harmonic_solver.py
@@ -153,7 +153,9 @@ class AcousticHarmonicSolver:
 
         # process the prescribed values
         self.prescribed_values, self.array_prescribed_values = self.assembler.get_prescribed_dofs_values()
+
         frequency_dependent = self.assembler.frequency_dependent
+
         for i, freq in enumerate(frequencies):
             logging.info(f"Solution step {i + 1} and frequency {freq} Hz [{i + 1}/{len(frequencies)}]")
             
@@ -172,7 +174,7 @@ class AcousticHarmonicSolver:
                 f_Qms = self.assembler.compute_mass_source_load_vector(1)
 
                 # compute the prescribed dofs-related load vector
-                f_eq = self.get_prescribed_pressure_model_excitation()
+                f_eq = self.assembler.get_prescribed_pressure_model_excitation(self.array_prescribed_values)
 
                 # compute the load vector f for omega = 1
                 f = f_Qms - 1j * f_Q[:, i] - f_eq
@@ -215,7 +217,7 @@ class AcousticHarmonicSolver:
                     self.assembler.assemble_mass_source_matrices_from_volumes(index=i)
 
                 # update the prescribed dofs-related load vector for each frequency step
-                f_eq = self.get_prescribed_pressure_model_excitation(index=i)
+                f_eq = self.assembler.get_prescribed_pressure_model_excitation(self.array_prescribed_values, index=i)
 
             # compute the mass source load vector
             f_Qms = self.assembler.compute_mass_source_load_vector(omega, index=i)
@@ -294,586 +296,3 @@ class AcousticHarmonicSolver:
             full_solution[self.prescribed_indexes] = self.array_prescribed_values[:, freq_index]
 
         return full_solution
-
-    def get_prescribed_pressure_model_excitation(self, index: int = 0):
-        """
-        This method computes the equivalent loads resulting from the degrees of freedom 
-        prescription to compound the acoustic model excitation vector.
-
-        Parameters
-        ----------
-        index: int, optional
-            An integer values that represents the frequency index.
-
-        Returns
-        -------
-        F_eq: np.ndarray
-            The equivalent acoustic load vector of complex numbers in which
-            each column corresponds to a frequency step of analysis.
-        """
-
-        if len(self.prescribed_values) == 0:
-            return 0.
-
-        frequencies = self.assembler.model.frequencies
-        omega = 2 * np.pi * frequencies[index]
-
-        values = self.array_prescribed_values[:, index]
-
-        self.Kr = self.assembler.stiffness_matrix_r
-        self.Mr = self.assembler.mass_matrix_r
-        self.Cr = self.assembler.damping_matrix_r
-        self.Cr_visc = self.assembler.visc_damping_matrix_r
-
-        Kr_add = self.Kr @ values
-        Mr_add = self.Mr @ values
-        Cr_add = (self.Cr + self.Cr_visc) @ values
-
-        F_Kadd = Kr_add
-        F_Madd = -(omega**2) * Mr_add 
-        F_Cadd = 1j * omega * Cr_add
-        F_eq = F_Kadd + F_Madd + F_Cadd
-
-        return F_eq[self.unprescribed_indexes]
-
-
-    def get_particle_velocity_from_surface(self, surface_id: int, rho: float | np.ndarray):
-        """ 
-        This method computes the nodal average particle velocity in the selected surface.
-
-        Parameters
-        ----------
-        surface_id: int
-            The selected surface ID.
-
-        rho: float
-            The fluid density related to the selected surface.
-        
-        Returns
-        -------
-
-        particle_velocities: dict
-            A dictionary with the normal particle velocity and its components in
-            the x, y, and z directions, computed in the selected surface.
-        """
-        # t0 = time()
-
-        frequencies = self.assembler.model.frequencies
-        element_3d = self.assembler.model.acoustic_element_3d
-
-        if element_3d is None:
-            self.assembler.define_acoustic_elements()
-            element_3d = self.assembler.element_3d
-
-        if element_3d.connectivity is None:
-            element_3d.reorder_connect()
-
-        data_normals = self.assembler.model.mesh.get_average_normals_for_surface_nodes(surface_id)
-        map_elements_to_nodes, filtered_nodes = self.assembler.model.mesh.get_solid_elements_connected_to_nodes(
-                                                                                                                surface_id = surface_id, 
-                                                                                                                return_nodes = True
-                                                                                                                )
-
-        # Load all frequency solutions to optimize multiple load on the `process_particle_velocity` method below.
-        node_to_index = dict(zip(filtered_nodes, np.arange(filtered_nodes.size, dtype=int)))
-        solution = self.solution[filtered_nodes, :]
-        # nodal_pressures = self.solution[:, :]
-
-        pv_data = dict()
-        for node_id, solid_element_ids in map_elements_to_nodes.items():
-
-            Vk = 0.
-            for element_id in solid_element_ids:
-                connect = element_3d.connectivity[element_id, 1:]
-                indexes = np.array([node_to_index.get(node) for node in connect])
-                enodal_pressures = solution[indexes, :]
-
-                Vk += element_3d.process_particle_velocity(
-                    element_id, 
-                    node_id, 
-                    rho, 
-                    frequencies,
-                    nodal_pressures = enodal_pressures,
-                    solution = None,
-                )
-
-            pv_data[node_id] = Vk / len(solid_element_ids)
-
-        Vx = dict()
-        Vy = dict()
-        Vz = dict()
-        Vn = dict()
-        particle_velocities = dict()
-
-        for i, _node_id in enumerate(np.sort(list(pv_data.keys()))):
-            Vx[_node_id] = pv_data[_node_id][0, :]
-            Vy[_node_id] = pv_data[_node_id][1, :]
-            Vz[_node_id] = pv_data[_node_id][2, :]
-            Vn[_node_id] = pv_data[_node_id].T @ data_normals[_node_id]
-
-        particle_velocities["Vx"] = Vx
-        particle_velocities["Vy"] = Vy
-        particle_velocities["Vz"] = Vz
-        particle_velocities["Vn"] = Vn
-        particle_velocities["nodal_normals"] = data_normals
-
-        ## Uncomment the line below to plot the average normals at the nodes
-        # self.assembler.model.mesh.set_nodal_normals_data(data_normals)
-
-        ## Only for validation purposes
-        # output_data = np.zeros((len(ordered_nodes), 4), dtype=float)
-        # output_data[:, 0] = ordered_nodes
-
-        # for row, node_id in enumerate(ordered_nodes):
-        #     output_data[row, 1:] =  self.assembler.model.mesh.nodal_normals_data[node_id]
-
-        # fname = f"nodal_normals_data_surface_{surface_id}.dat"
-        # header = "Node index || x-axis component [m] || y-axis component [m] || z-axis component [m]"
-        # np.savetxt(fname, output_data, fmt=["%i", "%.16f", "%.16f", "%.16f"], delimiter=",", header=header)
-
-        # dt = time() - t0
-        # print(f"Elpased time to get_particle_velocity_from_surface: {dt} s")
-
-        return particle_velocities
-
-
-    def get_transmission_loss(self, input_surface_id: int, output_surface_id: int, surface_integration: bool = True):
-        """ 
-        This method compute the acoustic transmission loss between two selected surfaces.
-
-        Parameters
-        ----------
-        input_surface_id: int
-            The input surface ID.
-
-        output_surface_id: int
-            The output surface ID.
-
-        surface_integration: bool, optional
-            It controls whether the sound power will be calculated using surface integration
-            or through the summation of the product of nodal sound intensity and nodal area.
-
-        Returns
-        -------
-        frequencies: np.ndarray
-            The vector of valid frequencies.
-
-        transmission_loss: np.ndarray
-            The vector of computed transmission loss values in dB.
-
-        """
-
-        model = self.assembler.model
-        frequencies = self.assembler.model.frequencies
-
-        nodes_input = np.sort(model.mesh.get_nodes_from_surface(input_surface_id))
-        nodes_output = np.sort(model.mesh.get_nodes_from_surface(output_surface_id))
-
-        P_in = self.solution[nodes_input, :]
-        P_out = self.solution[nodes_output, :]
-
-        logging.info("Processing the transmission loss... [40/100]")
-
-        if not surface_integration:
-
-            A_in = model.mesh.surface_area_from_element_integration.get(input_surface_id)
-            A_out = model.mesh.surface_area_from_element_integration.get(output_surface_id)
-
-            # print(f"A_in: {A_in} [m²]")
-            # print(f"A_out: {A_out} [m²]")
-
-            nodal_areas_in = np.zeros(len(nodes_input), dtype=float)
-            for i, node in enumerate(nodes_input):
-                areas = model.mesh.nodal_area[node]
-                nodal_areas_in[i] = sum(areas)
-
-            # _nodal_areas_in = np.array([nodes_input, nodal_areas_in * (A_in / np.sum(nodal_areas_in))]).T
-            # np.savetxt(f"nodal_areas_surface_{input_surface_id}.dat", _nodal_areas_in, fmt=["%i", "%.16f"], delimiter=",", header="Node index || Nodal area [m2]")
-
-            nodal_areas_out = np.zeros(len(nodes_output), dtype=float)
-            for i, node in enumerate(nodes_output):
-                areas = model.mesh.nodal_area[node]
-                nodal_areas_out[i] = sum(areas)
-
-            # _nodal_areas_out = np.array([nodes_output, nodal_areas_out * (A_out / np.sum(nodal_areas_out))]).T
-            # # np.savetxt(f"nodal_areas_surface_{output_surface_id}.dat", _nodal_areas_out, fmt=["%i", "%.16f"], delimiter=",", header="Node index || Nodal area [m2]")
-
-            Aeff_in = nodal_areas_in.reshape(-1, 1) * (A_in / np.sum(nodal_areas_in))
-            Aeff_out = nodal_areas_out.reshape(-1, 1) * (A_out / np.sum(nodal_areas_out))
-
-            rho_in, _ = model.get_fluid_properties_from_surface(input_surface_id, frequencies)
-            if rho_in is None:
-                return None, None
-
-            rho_out, _ = model.get_fluid_properties_from_surface(output_surface_id, frequencies)
-            if rho_out is None:
-                return None, None
-
-            logging.info("Processing the transmission loss... [50/100]")
-            input_pv_data = self.get_particle_velocity_from_surface(input_surface_id, rho_in)
-
-            logging.info("Processing the transmission loss... [60/100]")
-            output_pv_data = self.get_particle_velocity_from_surface(output_surface_id, rho_out)
-
-        logging.info("Processing the transmission loss... [70/100]")
-        Zo_in = self.get_surface_impedance(input_surface_id)
-        if Zo_in is None:
-            return None, None
-
-        Zo_out = self.get_surface_impedance(output_surface_id)
-        if Zo_out is None:
-            return None, None
-
-        logging.info("Processing the transmission loss... [80/100]")
-        (P_downstream, V_downstream) = self.get_downstream_pressure_and_particle_velocity(input_surface_id)
-        if P_downstream is None or V_downstream is None:
-            return None, None
-
-        logging.info("Processing the transmission loss... [90/100]")
-        if surface_integration: 
-            ## compute the sound power through surface integration
-            W_in = self.integrate_surface_sound_power(input_surface_id, P_downstream, np.conj(P_downstream / Zo_in))
-            W_out = self.integrate_surface_sound_power(output_surface_id, P_out, np.conj(P_out / Zo_out))
-
-        else:
-            ## compute the sound power through summation of the product of nodal sound intensity by nodal area
-            # input sound intensity calculation
-            I_in = np.abs(np.real(P_downstream * np.conjugate(V_downstream)) / 2)
-
-            # output sound intensity calculation
-            V_out = np.array(list(output_pv_data["Vn"].values()), dtype=complex)
-            I_out = np.real(P_out * np.conjugate(V_out)) / 2
-
-            # NOTE: be careful of using the calculated particle velocity in the sound power 
-            # integration. If the element shape functions are linear, the particle velocity
-            # will be constant in the element. This results is not consitent with the physics,
-            # therefore, you should use 'richer' elements with the quadratic shape functions 
-            # to get more representative results.
-
-            # V_in = -np.array(list(input_pv_data["Vn"].values()), dtype=complex)
-            # P_downstream = (P_in + Zo_in * V_in) / 2
-            # V_downstream = P_downstream / Zo_in
-
-            # I_in = np.real(P_downstream * np.conjugate(V_downstream)) / 2
-            # I_out = np.real(P_out * np.conjugate(V_out)) / 2
-
-            # compute the transmission loss using nodal areas
-            W_in = 10 * np.log10(np.sum(I_in * Aeff_in, axis=0))
-            W_out = 10 * np.log10(np.sum(I_out * Aeff_out, axis=0))
-
-            # compute the transmission loss using an alternative surface integration
-            # W_in = self.integrate_surface_sound_power_from_nodal_sound_intensity(input_surface_id, I_in)
-            # W_out = self.integrate_surface_sound_power_from_nodal_sound_intensity(output_surface_id, I_out)
-
-        transmission_loss = W_in - W_out
-
-        if frequencies[0] == 0:
-            frequencies = frequencies[1:]
-            transmission_loss = transmission_loss[1:]
-
-        return frequencies, transmission_loss
-
-
-    def get_surface_impedance(self, surface_id: int) -> float | complex | np.ndarray:
-        """
-        It returs the acoustic impedance of selected surface.
-
-        Parameter
-        ---------
-        surface_id: int
-            The selected surface ID.
-
-        Returns
-        -------
-        impedance: np.ndarray, float or None
-            The acoustic impedance of selected surface.
-        """
-
-        impedance = None
-        model = self.assembler.model
-
-        si_data = model.properties._get_property("specific_impedance", surface=surface_id)
-        pw_data = model.properties._get_property("incident_plane_wave", surface=surface_id)
-
-        if isinstance(si_data, dict):
-            if "real_values" in si_data.keys():
-                real_values = np.array(si_data["real_values"])
-                imag_values = np.array(si_data["imag_values"])
-                impedance = real_values + 1j * imag_values
-
-            elif "anechoic_termination" in si_data.keys():
-                rho_eff_pm, C_eff_pm = model.get_porous_material_model_effective_properties(surface_id)
-                rho_eff_tv, C_eff_tv = model.get_viscous_thermal_model_effective_properties(surface_id)
-
-                if isinstance(rho_eff_pm, np.ndarray):
-                    density = rho_eff_pm
-                    speed_of_sound = C_eff_pm
-
-                elif isinstance(rho_eff_tv, np.ndarray):
-                    density = rho_eff_tv
-                    speed_of_sound = C_eff_tv
-
-                else:
-
-                    fluid = model.properties._get_property("fluid", surface=surface_id)
-                    if not isinstance(fluid, Fluid):
-                        return None
-
-                    density = fluid.fluid_density
-                    speed_of_sound = fluid.speed_of_sound
-
-                impedance = density * speed_of_sound
-
-            elif "values" in si_data.keys():
-                impedance = si_data["values"][0]
-
-        elif isinstance(pw_data, dict):
-            rho_eff_pm, C_eff_pm = model.get_porous_material_model_effective_properties(surface_id)
-            rho_eff_tv, C_eff_tv = model.get_viscous_thermal_model_effective_properties(surface_id)
-
-            if isinstance(rho_eff_pm, np.ndarray):
-                density = rho_eff_pm
-                speed_of_sound = C_eff_pm
-
-            elif isinstance(rho_eff_tv, np.ndarray):
-                density = rho_eff_tv
-                speed_of_sound = C_eff_tv
-
-            else:
-                fluid = model.properties._get_property("fluid", surface=surface_id)
-                if not isinstance(fluid, Fluid):
-                    return None
-
-                density = fluid.fluid_density
-                speed_of_sound = fluid.speed_of_sound
-
-            impedance = density * speed_of_sound
-
-        return impedance
-
-
-    def get_downstream_pressure_and_particle_velocity(self, surface_id: int):
-        """
-        This method computes the downstream pressure and particle velocity
-        from the model acoustic excitation.
-
-        Parameters
-        ----------
-        surface_id: int
-            The input surface ID.
-
-        Returns
-        -------
-        P_downstream: np.ndarray
-            The downstream pressure vector or matrix.
-
-        V_downstream: np.ndarray
-            The downstream velocity vector or matrix.
-        """
-
-        model = self.assembler.model
-        frequencies = self.assembler.model.frequencies
-
-        Zo_in = self.get_surface_impedance(surface_id)
-        if Zo_in is None:
-            return None, None
-
-        pw_data = model.properties._get_property("incident_plane_wave", surface=surface_id)
-        sv_data = model.properties._get_property("surface_velocity", surface=surface_id)
-
-        if not (pw_data or sv_data):
-            return None, None
-
-        if isinstance(pw_data, dict):
-            values = pw_data.get("values")[0]
-            _wave_vector = pw_data.get("wave_vector")
-            wave_vector = np.array(_wave_vector, dtype=float)
-
-            if isinstance(values, complex | float):
-                P_inc = values * np.ones_like(frequencies, dtype=complex)
-            else:
-                P_inc = values
-
-            node_normals = model.mesh.get_stacked_normals_for_surface_elements(surface_id)
-            avg_normal = np.average(node_normals, axis=0).flatten()
-
-            P_downstream = P_inc * (avg_normal @ wave_vector)
-            V_downstream = -P_downstream / Zo_in
-
-        if isinstance(sv_data, dict):
-            if "real_values" in sv_data.keys():
-                real_values = np.array(sv_data["real_values"])
-                imag_values = np.array(sv_data["imag_values"])
-                V_in = real_values + 1j * imag_values
-
-            elif "values" in sv_data.keys():
-                V_in = sv_data["values"]
-
-            P_downstream = V_in * Zo_in / 2
-            V_downstream = P_downstream / Zo_in
-
-        return P_downstream, V_downstream
-
-
-    def integrate_surface_sound_power(  
-                                      self, 
-                                      surface_id: int,
-                                      pressures: np.ndarray,
-                                      particle_velocities: np.ndarray,
-                                      dB_scale: bool = True
-                                      ) -> np.ndarray:
-        """
-        This method integrates the sound power intensity over the selected surface.
-
-        Parameters
-        ----------
-            surface_id: int
-                The identifier of selected surface.
-
-            pressures: np.ndarray
-                The acoustic pressures from selected suraface.
-
-            particle_velocities: np.ndarray
-                The acoustic pressures from selected suraface.
-
-        Returns
-        -------
-        sound_power: np.ndarray
-            The sound power level in dB if dB_scale is True or the sound power in watts otherwise.
-        """
-
-        nodes = np.sort(self.assembler.model.mesh.get_nodes_from_surface(surface_id))
-        surface_connectivities = self.assembler.model.mesh.get_connectivity_from_surface(surface_id)
-
-        number_nodes = len(nodes)
-        map_nodes = dict(zip(nodes, np.arange(number_nodes)))
-
-        if len(pressures.shape) == 1:
-            pressures = np.tile(pressures, (number_nodes, 1))
-
-        if len(particle_velocities.shape) == 1:
-            particle_velocities = np.tile(particle_velocities, (number_nodes, 1))
-
-        element_2d = self.assembler.element_2d
-        if element_2d is None:
-            self.assembler.define_acoustic_elements()
-            element_2d = self.assembler.element_2d
-
-        sound_power = 0.
-        for i, e_connect in enumerate(surface_connectivities):
-            node_indexes = [map_nodes.get(node) for node in e_connect]
-            L_sv = pressures[node_indexes, :].T.reshape(-1, 1, 3)
-            R_sv = particle_velocities[node_indexes, :].T.reshape(-1, 3, 1)
-
-            normalized_data = element_2d.elementary_sound_power(e_connect, L_sv, R_sv)
-            sound_power += np.real(normalized_data) / 2
-
-        if dB_scale:
-            return 10 * np.log10(sound_power / 1e-12)
-
-        return sound_power
-
-
-    def integrate_surface_sound_power_from_nodal_sound_intensity(  
-                                                                 self, 
-                                                                 surface_id: int,
-                                                                 sound_intensities: np.ndarray,
-                                                                 dB_scale: bool = True
-                                                                 ) -> np.ndarray:
-        """
-        This method integrates the sound power intensity over the selected surface.
-
-        Parameters
-        ----------
-            surface_id: int
-                The identifier of selected surface.
-
-            sound_intensities: np.ndarray
-                The acoustic sound intensities from selected suraface.
-
-        Returns
-        -------
-        sound_power: np.ndarray
-            The sound power level in dB if dB_scale is True or the sound power in watts otherwise.
-        """
-
-        nodes = np.sort(self.assembler.model.mesh.get_nodes_from_surface(surface_id))
-        surface_connectivities = self.assembler.model.mesh.get_connectivity_from_surface(surface_id)
-
-        number_nodes = len(nodes)
-        map_nodes = dict(zip(nodes, np.arange(number_nodes)))
-
-        if len(sound_intensities.shape) == 1:
-            sound_intensities = np.tile(sound_intensities, (number_nodes, 1))
-
-        element_2d = self.assembler.element_2d
-        if element_2d is None:
-            self.assembler.define_acoustic_elements()
-            element_2d = self.assembler.element_2d
-
-        sound_power = 0.
-        for i, e_connect in enumerate(surface_connectivities):
-            node_indexes = [map_nodes.get(node) for node in e_connect]
-            sound_intensity = sound_intensities[node_indexes, :]
-            normalized_data = element_2d.elementary_sound_power_from_sound_intensity(e_connect, sound_intensity)
-            sound_power += np.sum(np.real(normalized_data) / 2, axis=0)
-
-        if dB_scale:
-            return 10 * np.log10(sound_power / 1e-12)
-
-        return sound_power
-
-
-    def get_noise_reduction(self, input_surface_id, output_surface_id):
-        """ 
-        This method compute the acoustic noise reduction between two selected surfaces.
-
-        Parameters
-        ----------
-
-        input_surface_id: int
-            The input surface ID.
-
-        output_surface_id: int
-            The output surface ID.
-
-        Returns
-        -------
-
-        frequencies: np.ndarray
-            The vector of valid frequencies.
-
-        noise_reduction: np.ndarray
-            The vector of computed noise reduction values in dB.
-
-        """
-        frequencies = self.assembler.model.frequencies
-        rows_input = self.assembler.model.mesh.get_nodes_from_surface(input_surface_id)
-        rows_output = self.assembler.model.mesh.get_nodes_from_surface(output_surface_id)
-
-        P_in = np.average(self.solution[rows_input,:], axis=0)
-        P_out = np.average(self.solution[rows_output,:], axis=0)
-
-        # the zero_shift constant is summed to avoid zero values either in P_input2 or P_output2 variables
-        zero_shift = 1e-12
-
-        Prms_out2 = np.real(P_out*np.conjugate(P_out)) / 2 + zero_shift
-        Prms_in2 = np.real(P_in*np.conjugate(P_in)) / 2 + zero_shift
-
-        noise_reduction = 10*np.log10(Prms_in2/Prms_out2)
-
-        if frequencies[0] == 0:
-            frequencies = frequencies[1:]
-            noise_reduction = noise_reduction[1:]
-
-        return frequencies, noise_reduction
-
-
-    def plot_graph(self, matrix):
-        """
-        """
-        import matplotlib.pyplot as plt
-        plt.ion()
-        plt.cla()
-        plt.spy(matrix, color=(0.25,0.25,0.25))
-        plt.show()
